bots/trendobot/requirements.txt
asyncio==3.4.3
certifi==2025.1.31
matplotlib
nats-py
numpy
pandas
pandas_ta
pyarrow
PyYAML
requests==2.32.3
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
TA-Lib
tzdata==2025.1
upstox-python-sdk
urllib3==2.3.0
websocket-client==1.8.0
websockets==15.0
